// src/store/main/analysis/types.ts
/*
 * @Author: tcosfish
 * @Date: 2022-06-14 14:23:06
 * @LastEditors: tcosfish
 * @LastEditTime: 2022-06-14 15:10:33
 * @FilePath: \vue3admin\src\store\main\analysis\types.ts
 */
export interface IAnalysisState {
  categoryGoodsCount: any[];
  categoryGoodsSale: any[];
  categoryGoodsFavor: any[];
  addressGoodsSale: any[];
}